Enhance CMake configuration, add dtype tests, and improve tensor functionality
- Updated CMakeLists.txt for library installation and export configuration.
- Added new methods to TensorExtents for better range access.
- Refactored Dtype::from() for improved type handling.
- Expanded Tensor class methods for better data access.
- Introduced TensorOptions constructors for flexibility.
- Added unit tests for Dtype functionality.
- Improved image loading and saving functions.
- Updated vcpkg.json with enhanced metadata and dependencies.

+++ native/cpp/io/lib/cnpy/cnpy.h
@@ -164,7 +164,7 @@ namespace cnpy {
 
         std::vector<char> npy_header = create_npy_header<T>(shape);
 
-        size_t nels = std::accumulate(shape.begin(),shape.end(),1,std::multiplies<size_t>());
+        size_t nels = std::accumulate(shape.begin(),shape.end(),size_t(1),std::multiplies<size_t>());
         size_t nbytes = nels*sizeof(T) + npy_header.size();
 
         //get the CRC of the data to be added
